Foi realizado o refactory nos serializadores assim como foi feito a adequação onde eles era utilizadas
closes #18
