+++ app/api.http
@@ -10,32 +10,32 @@ POST {{baseURL}}/api/states
 Content-Type: application/json
 
 {
-    "name": "Pará",
-    "abbreviation": "PA"
+    "name": "Naranhão",
+    "abbreviation": "MA"
 }
 
 ###
 # @name getOneState
-GET {{baseURL}}/api/states/1
+GET {{baseURL}}/api/states/8
 
 ###
 # @name deleteOneState
-DELETE {{baseURL}}/api/states/6
+DELETE {{baseURL}}/api/states/7
 
 
 ###
 # @name updateOneState
-PUT {{baseURL}}/api/states/22
+PUT {{baseURL}}/api/states/9
 Content-Type: application/json
 
 {
-    "name": "Acre",
-    "abbreviation": "TO"
+    "name": "Maranhão",
+    "abbreviation": "MA"
 }
 
 ###
 # @name allCities
-GET {{baseURL}}/api/cities?filters=[{"property": "name__icontains", "value": "flor", "stage": 1}]
+GET {{baseURL}}/api/cities
 
 
 ###
@@ -44,17 +44,17 @@ POST {{baseURL}}/api/cities
 Content-Type: application/json
 
 {
-    "name": "Divinopolis",
-    "state_id": 1
+    "name": "Balsas",
+    "state_id": 9
 }
 
 ###
 # @name getoneCity
-GET {{baseURL}}/api/cities/1
+GET {{baseURL}}/api/cities/5
 
 ###
 # @name updateOneCity
-PUT {{baseURL}}/api/cities/1
+PUT {{baseURL}}/api/cities/5
 Content-Type: application/json
 
 {
@@ -63,7 +63,7 @@ Content-Type: application/json
 
 ###
 # @name deleteOneCity
-DELETE {{baseURL}}/api/cities/1
+DELETE {{baseURL}}/api/cities/5
